Cache result to stem memory leak
If called repeatedly it would leak lots of memory.

Some of that was from the sub created in the EVAL and it appears
some of it comes from the failed native call.

The one drawback is it won't detect the sudden appearance of the
library, but how often does that happen.

Fixes #3